requirements.txt
flask
psycopg[binary]
sqlalchemy
sqlalchemy-cockroachdb